클라이언트용 서버 모듈의 설정 경로 수정
