+++ config0.yaml
@@ -1,5 +1,6 @@
-address: 'localhost'
-port: 2421
+client:
+  address: 'localhost'
+  port: 2421
 cluster:
   address: 'localhost'
   port: 9090
