+++ src/main/java/databute/databuter/DatabuterConfiguration.java
@@ -3,26 +3,20 @@ package databute.databuter;
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.base.MoreObjects;
+import databute.databuter.client.ClientConfiguration;
 import databute.databuter.cluster.ClusterConfiguration;
 
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY)
 public class DatabuterConfiguration {
 
-    @JsonProperty("address")
-    private String address;
-
-    @JsonProperty("port")
-    private int port;
+    @JsonProperty("client")
+    private ClientConfiguration client;
 
     @JsonProperty("cluster")
     private ClusterConfiguration cluster;
 
-    public String address() {
-        return address;
-    }
-
-    public int port() {
-        return port;
+    public ClientConfiguration client() {
+        return client;
     }
 
     public ClusterConfiguration cluster() {
@@ -32,8 +26,7 @@ public class DatabuterConfiguration {
     @Override
     public String toString() {
         return MoreObjects.toStringHelper(this)
-                .add("address", address)
-                .add("port", port)
+                .add("client", client)
                 .add("cluster", cluster)
                 .toString();
     }
